refactor formStore by creating store folder
