+++ src/constants/index.js
@@ -1,0 +1,4 @@
+export * from './character';
+export * from './contact';
+export * from './skills';
+export * from './text';
